+++ tests/client/mod.rs
@@ -1,2 +1,3 @@
 mod completions;
+mod log;
 mod status;