+++ client/display/state.rs
@@ -157,6 +157,11 @@ fn print_table(tasks: &[Task], style: &OutputStyle, settings: &Settings) {
         .load_preset(UTF8_HORIZONTAL_ONLY)
         .set_header(headers);
 
+    // Explicitly force styling, in case we aren't on a tty, but `--color=always` is set.
+    if style.enabled {
+        table.enforce_styling();
+    }
+
     // Add rows one by one.
     for task in tasks.iter() {
         let mut row = Row::new();
@@ -186,7 +191,7 @@ fn print_table(tasks: &[Task], style: &OutputStyle, settings: &Settings) {
                 enqueue_at: Some(enqueue_at),
             } = task.status
             {
-                // Only show the date if the task is supposed to be enqueued today.
+                // Only show the date if the task is not supposed to be enqueued today.
                 let enqueue_today =
                     enqueue_at <= Local::today().and_hms(0, 0, 0) + Duration::days(1);
                 let formatted_enqueue_at = if enqueue_today {