+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+ ## Unreleased
+
+ ### Fixed
+
+ - Only style the 'group' header in status output when on a TTY.
 
 ## [2.0.4] - 2022-06-05